Added augmentation to neural network models.
