+++ nitorch/nn/models/__init__.py
@@ -1,0 +1,3 @@
+"""Models (= task-specific models)"""
+
+from .segmentation import *
