+++ nitorch/nn/modules/__init__.py
@@ -5,5 +5,4 @@ from ._cnn import *
 from ._conv import *
 from ._norm import *
 from ._pool import *
-from ._segmentation import *
 from ._spatial import *
